fixup tests
